feat(neovim): replace NvChad with custom config
NvChad is great but it can be troublesome to configure, besides I can't
event use half of the features it provides since I don't know the
keymaps assigned to them. That's why I'm making my own config where I
define all the keymaps and add features, so hopefully I would know how
to use it.

WIP: This config is not yet completed, there is a lot of keymaps missing
and features to configure.